% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example cohort structure.}
\format{A data frame with six variables: \code{Batch}, \code{ID},
  \code{DiseaseType}, \code{Gender} and \code{Patient} (\code{Time}}
\usage{
example_data
}
\description{
Sample dataset summarising a plausible cohort structure for a set of samples available for patients over time, across different batches and
with different disease outcomes
}
\keyword{datasets}
