% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\title{Example matadata.}
\format{A data frame with six variables: \code{SampleName}, \code{PID},\code{Time},
\code{Blast}, \code{Outcome}}
\usage{
example_metadata
}
\description{
Sample dataset summarising a plausible metadata file for a set of samples available for patients over time
}
\keyword{datasets}
