% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_variants}
\alias{example_variants}
\title{Example variants.}
\format{A data frame with six variables: \code{SampleName}, \code{PID},\code{Time},
\code{mutation_key}, \code{chrom},\code{pos},
\code{alt}, \code{ref},\code{ref_depth}, \code{alt_depth},
\code{VAF}, \code{mutation_det},\code{SYMBOL}, \code{Consequence}}
\usage{
example_variants
}
\description{
Sample dataset summarising a plasible set of variants called across two patients
}
\keyword{datasets}
