% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_clones.R
\name{import_clones_for_lineplot}
\alias{import_clones_for_lineplot}
\title{Import clones detected by superFreq for lineplots}
\usage{
import_clones_for_lineplot(stories_dir, patientID)
}
\arguments{
\item{stories_dir}{path to where the superFreq stories.Rdata files is saved.}

\item{patientID}{a character vector specifying the patient/s id/s for which stories have to be imported.}
}
\description{
Import clones detected by superFreq for lineplots
}
