% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cnas.R
\name{import_cnas}
\alias{import_cnas}
\title{Import copy number variation detected by superFreq for a specific patient}
\usage{
import_cnas(mutationSummary = NA, patientID = NA, tidy = TRUE,
  time_order = c("Screen", "Cyc1", "Cyc2", "Cyc3", "Cyc4", "Cyc9"),
  sample_name_parts = c("PID", "Time", "Status", "Repl.Within", "Batch",
  "Outcome"), ref_genome = "hg38")
}
\arguments{
\item{mutationSummary}{a matrix summarising mutations reported by superFreq. Every row is a CN event, every column is a sample and every entry is the VAF for that row mutation for that column sample.}

\item{patientID}{a character vector specifying the patient/s id/s for which mutations have to be imported. This value has to correspond to the names used to save the `mutationSummary` object.}

\item{tidy}{Logical. Should the ouput be in a tidy or untidy (list of matrices) format? Default is `tidy = TRUE`.}

\item{time_order}{vector specifying the order of time points, e.g. the levels of the `Time` variable `sample_name_parts` definition. This will be used to order samples for plotting purposes.}

\item{sample_name_parts}{vector with constitutive parts of the names for every sample. For example, if the sample names are pf the format: patientID.Time.Status.Repl.within.batch.Outcome, the vector will be sample_name_parts = c("patientID","Time","Status","Repl.Within","Batch","Outcome"). The component `Time` is required.}

\item{ref_genome}{character specifying the reference genome used. Default is 'hg38'.}
}
\description{
Import copy number variation detected by superFreq for a specific patient
}
