% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_snvs.R
\name{import_snvs_for_lineplot}
\alias{import_snvs_for_lineplot}
\title{Import point mutations for lineplots}
\usage{
import_snvs_for_lineplot(mutationSummary, patientID, tidy = TRUE)
}
\arguments{
\item{mutationSummary}{a matrix summarising mutations reported by superFreq. Every row is a muations, every column is a sample and every entry is the VAF for that row mutation for that column sample.}

\item{patientID}{a character vector specifying the patient/s id/s for which mutations have to be imported. This value has to correspond to the names used to save the `mutationSummary` object.}

\item{tidy}{Logical. Should the ouput be in a tidy or untidy (list of matrices) format? Default is `tidy = TRUE`.}
}
\description{
Import point mutations for lineplots
}
\examples{

mutationSummary <- list()
mutationSummary[[1]] <- data.frame(x1 = c(1000,2000),
x2 = c(1000,2001),
call = c("1001+TCCG","chr22.A"),
stories.D1.Screen.Diag.R1.B1.Rel = c(0.33,0.74),
stories.D1.Cyc1.Rem.R1.B1.Rel = c(0,0.68),
stories.D1.Cyc2.Rem.R1.B1.Rel = c(0,0.1),
errors.D1.Screen.Diag.R1.B1.Rel = c(0,0.5),
errors.D1.Cyc1.Rem.R1.B1.Rel = c(0.04, 0.2),
errors.D1.Cyc2.Rem.R1.B1.Rel = c(0.2,0.1),
GoI = c("FLT3","STAG2"),
label = c("FLT3 (13) frameshift","1Mb A"),
severity = c(5,0),
chr = c(13,22),
start = c(2029,4554),
end = c(2035,4600))
rownames(mutationSummary[[1]]) <- mutationSummary[[1]]$call
names(mutationSummary) <- "D1"

}
