% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_vaf_varscan}
\alias{parse_vaf_varscan}
\title{Functions not to be exported}
\usage{
parse_vaf_varscan(freq)
}
\arguments{
\item{freq}{Vector of frequencies as recorded in VarScan VCF file, field FREQ.}
}
\value{
numeric vector of variant allele frequecies
}
\description{
Functions not to be exported
}
\keyword{internal}
