% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vep_csq.R
\name{parse_vep_csq}
\alias{parse_vep_csq}
\title{Parse the CSQ field added to the `VCF` file by the Variant Effect Predictor (VEP)}
\usage{
parse_vep_csq(vcf_path, vcf_df)
}
\arguments{
\item{vcf_path}{path to where the `.vcf` file for one sample is saved.}

\item{vcf_df}{`R` object returned by `parse_vcf_output()` after parsing relevant fields from the `VCF` file.}
}
\value{
data frame with VEP annoation fields added as columns.
}
\description{
Parse the CSQ field added to the `VCF` file by the Variant Effect Predictor (VEP)
}
