% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mutation_lines.R
\name{plot_mutation_lines}
\alias{plot_mutation_lines}
\title{Function to plot the lineplots with all clinical and genomic information}
\usage{
plot_mutation_lines(blasts_to_plot, indels_to_plot, snvs_to_plot,
  cnas_to_plot, clones_to_plot, spread = 0.1)
}
\arguments{
\item{blasts_to_plot}{output from `import_blast_for_lineplots()`}

\item{indels_to_plot}{output from `import_indels_for_lineplots()`}

\item{snvs_to_plot}{output from `import_snvs_for_lineplots()`}

\item{cnas_to_plot}{output from `import_cnas_for_lineplots()`}

\item{clones_to_plot}{output from `import_clones_for_lineplots()`}

\item{spread}{spread}
}
\description{
Function to plot the lineplots with all clinical and genomic information
}
