% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_blast.R
\name{import_blast_for_lineplot}
\alias{import_blast_for_lineplot}
\title{Import clones detected by superFreq for lineplots}
\usage{
import_blast_for_lineplot(clinicalData, patientID)
}
\arguments{
\item{clinicalData}{data frame with the `Blast` information (as %) for the patients in the cohort.}

\item{patientID}{a character vector specifying the patient/s id/s for which stories have to be imported.}
}
\description{
Import clones detected by superFreq for lineplots
}
