% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_clinical.R
\name{import_clinical}
\alias{import_clinical}
\title{Import clinical information for a specific patient}
\usage{
import_clinical(clinicalData = NA, patientID = NA, tidy = TRUE,
  extract_column = "Blast")
}
\arguments{
\item{clinicalData}{data frame containing the clinical information for a patient. This needs to include a column `SampleName` containing unique names for all the RNA-Seq samples available in the cohort; a `PID` columns for patient ID and a `Time` column that should reflect the time of sample collection.}

\item{patientID}{a character vector specifying the patient id for which clinical data should be imported.}

\item{tidy}{Logical. Should the ouput be in a tidy or untidy (list of matrices) format? Default is `tidy = TRUE`.}

\item{extract_column}{vector specifying the column to be extracted. Currently only numeric variables with values between 0 and 1 are allowed to be consistent with plotting together with clonality and VAF estimates.}
}
\description{
Import clinical information for a specific patient
}
\examples{


clinicalData <- data.frame(SampleName = c("D1.Screen.Rel","D1.Cyc1.Rel",
                                          "D1.Cyc2.Rel","D1.Cyc3.Rel"),
                           AgeDiagnosis = 65,
                           Sex = "F",
                           Blast = c(80,5,7,40)/100) \%>\%
  tidyr::separate(SampleName,into=c("PID","Time","Outcome"),sep = "[.]",remove=FALSE)


import_blast <- import_clinical(tidy = TRUE,extract_column = "Blast",clinicalData = clinicalData,
                                patientID = "D1")

}
