% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_indels.R
\name{import_indels_for_lineplot}
\alias{import_indels_for_lineplot}
\title{Import INDELs detected by VarDict}
\usage{
import_indels_for_lineplot(indels = indels, patientID, studyGenes,
  minQual = 20, minLength = 2, clinicalData)
}
\arguments{
\item{indels}{a data frame where every row is an INDEL for one sample at a specific time point}

\item{patientID}{a character vector specifying the patient/s id/s for which INDELs have to be imported.}

\item{studyGenes}{genes of interest used to subset the `indels` data frame.}

\item{minQual}{minimum quality for an INDEL to be kept.}

\item{minLength}{minimum length of the alternative allele required for an INDEL to be kept.}

\item{clinicalData}{clinical data about the patients in the cohort. It has to contain a column `SampleName` which is made of the following information `OurPID`,`Time`,`Status`,`Repl.within`,`batch` and `Outcome` separated by a '.'.}
}
\description{
This function will take as input a data frame of INDELs `indels`. It will keep only the INDELs for `patientID` on `studyGenes` and with a `minQual` and `minLength` of the alt allele. It will return a matrix where the rows are all the unique INDELs found for `patientID` across time and the columns are the samples of `patientID` across time. Every entry of this matrix is the VAF for each sample for that row mutation. An INDEL is reported in output only if the max VAF across time is > 0.15
}
