% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_snvs.R
\name{import_snvs_for_lineplot}
\alias{import_snvs_for_lineplot}
\title{Import point mutations for lineplots}
\usage{
import_snvs_for_lineplot(mutationSummary, patientID)
}
\arguments{
\item{mutationSummary}{a matrix summarising mutations reported by superFreq. Every row is a muations, every column is a sample and every entry is the VAF for that row mutation for that column sample.}

\item{patientID}{a character vector specifying the patient/s id/s for which mutations have to be imported. This value has to correspond to the names used to save the `mutationSummary` object.}
}
\description{
Import point mutations for lineplots
}
