% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_vcf_output.R
\name{parse_vcf_output}
\alias{parse_vcf_output}
\title{Parse a VCF file and return a data frame with standardised fields across callers to use for caller comparison.}
\usage{
parse_vcf_output(vcf_path, sample_name, caller, vep = TRUE)
}
\arguments{
\item{vcf_path}{path to where the `.vcf` file for one sample is saved.}

\item{sample_name}{character. Sample name of the current `vcf` file.}

\item{caller}{character. One of `mutect`, `vardict` or `varscan`.}

\item{vep}{logical. If TRUE, the annotation fields added by the Variant Effect Predictor will be parsed.}
}
\description{
Currently, it doesn't work for somatic calls and for VCF from the following callers: GATK3 MuTect2, VarScan2 and VarDict. It used the Bioconductor package `VariantAnnotation` to read `VCF` files into `R`.
}
\examples{
vcf_path <- system.file("extdata", "germline_mutect.vcf", package = "varikondo")
parsed_vcf_mutect <- varikondo::parse_vcf_output(annot_vcf_mutect,
caller = "mutect",
sample_name = "Sample1",
vep = TRUE)
}
